adição alteração e visualização no sga de testemunho, reflexao, usuario e video

+++ dao/TestemunhoDAO.php
@@ -7,15 +7,23 @@
       }
 
       public function adicionar($testemunho){
+        $retorno = false;
+        try{
 
-        $sql = $this->db->prepare(
-          "INSERT INTO testemunho VALUES(null,:Usuario_idUsuario, :email, :descricao, :nome)"
-        );
-        $sql->bindValue(":email", $testemunho->getEmail());
-        $sql->bindValue(":descricao", $testemunho->getDescricao());
-        $sql->bindValue(":nome", $testemunho->getNome());
-        $sql->bindValue(":Usuario_idUsuario", unserialize($_SESSION['usuario'])->getId());
-        $sql->execute();
+          $sql = $this->db->prepare(
+            "INSERT INTO testemunho VALUES(null,:Usuario_idUsuario, :email, :descricao, :nome)"
+          );
+          $sql->bindValue(":email", $testemunho->getEmail());
+          $sql->bindValue(":descricao", $testemunho->getDescricao());
+          $sql->bindValue(":nome", $testemunho->getNome());
+          $sql->bindValue(":Usuario_idUsuario", unserialize($_SESSION['usuario'])->getId());
+          if($sql->execute()){
+            $retorno = true;
+          }
+        }catch(Exception $e){
+          throw $e->getMessage();
+        }
+        return $retorno;
 
       }
       public function getTestemunhos($page, $perPage){
@@ -65,18 +73,35 @@
         return null;
       }
   public function apagar($id){
-    $sql = $this->db->prepare("DELETE FROM testemunho WHERE idTestemunho=:id");
-    $sql->bindValue("id", $id);
-    $sql->execute();
+    $retorno = false;
+    try{
+        if(!empty($id)){
+          $sql = $this->db->prepare("DELETE FROM testemunho WHERE idTestemunho=:id");
+          $sql->bindValue("id", $id);
+          if($sql->execute()){
+            $retorno = true;
+          }
+        }
+      }catch(Exception $e){
+        throw $e->getMessage();
+      }
+    return $retorno;
   }
   public function alterar($testemunho){
-
-    $sql = $this->db->prepare("UPDATE testemunho SET email=:email, nome=:nome, descricao=:descricao WHERE idTestemunho = :id");
-    $sql->bindValue(":email", $testemunho->getEmail());
-    $sql->bindValue(":nome", $testemunho->getNome());
-    $sql->bindValue(":descricao", $testemunho->getDescricao());
-    $sql->bindValue(":id", $testemunho->getId());
-    $sql->execute();
+    $retorno = false;
+    try{
+      $sql = $this->db->prepare("UPDATE testemunho SET email=:email, nome=:nome, descricao=:descricao WHERE idTestemunho = :id");
+      $sql->bindValue(":email", $testemunho->getEmail());
+      $sql->bindValue(":nome", $testemunho->getNome());
+      $sql->bindValue(":descricao", $testemunho->getDescricao());
+      $sql->bindValue(":id", $testemunho->getId());
+      if($sql->execute()){
+        $retorno = true;
+      }
+    }catch(Exception $e){
+      throw $e->getMessage();
+    }
+    return $retorno;
   }
   public function getTotal(){
     $sql = $this->db->query("SELECT COUNT(*) as c FROM testemunho");
